Performance Improvement
Determine max key length in each conversion dictionary.
To save time, when looking for a match in a string, only use sub-strings
that are equal or less than the max possible length.
~25% speed improvement when converting 魯迅︰吶喊 from traditional to
simplified; verified output text identical to original code. Tested on
python 2.7 and 3.7.

Add tests to achieve 100% code coverage.